+++ src/RequestBuilder/RecommendationRequestBuilder.php
@@ -5,8 +5,8 @@ namespace Lmc\Matej\RequestBuilder;
 use Fig\Http\Message\RequestMethodInterface;
 use Lmc\Matej\Exception\LogicException;
 use Lmc\Matej\Model\Command\Interaction;
+use Lmc\Matej\Model\Command\UserItemRecommendation;
 use Lmc\Matej\Model\Command\UserMerge;
-use Lmc\Matej\Model\Command\UserRecommendation;
 use Lmc\Matej\Model\Request;
 use Lmc\Matej\Model\Response\RecommendationsResponse;
 
@@ -21,12 +21,12 @@ class RecommendationRequestBuilder extends AbstractRequestBuilder
     private $interactionCommand;
     /** @var UserMerge|null */
     private $userMergeCommand;
-    /** @var UserRecommendation */
-    private $userRecommendationCommand;
+    /** @var UserItemRecommendation */
+    private $userItemRecommendationCommand;
 
-    public function __construct(UserRecommendation $userRecommendationCommand)
+    public function __construct(UserItemRecommendation $userItemRecommendationCommand)
     {
-        $this->userRecommendationCommand = $userRecommendationCommand;
+        $this->userItemRecommendationCommand = $userItemRecommendationCommand;
     }
 
     /** @return $this */
@@ -53,7 +53,7 @@ class RecommendationRequestBuilder extends AbstractRequestBuilder
         return new Request(
             static::ENDPOINT_PATH,
             RequestMethodInterface::METHOD_POST,
-            [$this->interactionCommand, $this->userMergeCommand, $this->userRecommendationCommand],
+            [$this->interactionCommand, $this->userMergeCommand, $this->userItemRecommendationCommand],
             $this->requestId,
             RecommendationsResponse::class
         );
@@ -75,14 +75,14 @@ class RecommendationRequestBuilder extends AbstractRequestBuilder
         $interactionUserId = $this->interactionCommand->getUserId();
 
         // (A, null, A)
-        if ($this->userMergeCommand === null && $interactionUserId !== $this->userRecommendationCommand->getUserId()) {
-            throw LogicException::forInconsistentUserId($this->userRecommendationCommand, $this->interactionCommand);
+        if ($this->userMergeCommand === null && $interactionUserId !== $this->userItemRecommendationCommand->getUserId()) {
+            throw LogicException::forInconsistentUserId($this->userItemRecommendationCommand, $this->interactionCommand);
         }
 
         // allow (B, A -> B, B)
         if ($this->userMergeCommand !== null
             && $interactionUserId === $this->userMergeCommand->getUserId()
-            && $interactionUserId === $this->userRecommendationCommand->getUserId()) {
+            && $interactionUserId === $this->userItemRecommendationCommand->getUserId()) {
             return;
         }
 
@@ -102,8 +102,8 @@ class RecommendationRequestBuilder extends AbstractRequestBuilder
     private function assertUserMergeUserId(): void
     {
         if ($this->userMergeCommand !== null
-            && $this->userMergeCommand->getUserId() !== $this->userRecommendationCommand->getUserId()) {
-            throw LogicException::forInconsistentUserId($this->userRecommendationCommand, $this->userMergeCommand);
+            && $this->userMergeCommand->getUserId() !== $this->userItemRecommendationCommand->getUserId()) {
+            throw LogicException::forInconsistentUserId($this->userItemRecommendationCommand, $this->userMergeCommand);
         }
     }
 }