+++ src/RequestBuilder/RequestBuilderFactory.php
@@ -4,7 +4,7 @@ namespace Lmc\Matej\RequestBuilder;
 
 use Lmc\Matej\Http\RequestManager;
 use Lmc\Matej\Model\Command\Sorting;
-use Lmc\Matej\Model\Command\UserRecommendation;
+use Lmc\Matej\Model\Command\UserItemRecommendation;
 
 /**
  * Factory to create concrete RequestBuilder which helps you to create request for each Matej API
@@ -74,7 +74,7 @@ class RequestBuilderFactory
         return $requestBuilder;
     }
 
-    public function recommendation(UserRecommendation $recommendation): RecommendationRequestBuilder
+    public function recommendation(UserItemRecommendation $recommendation): RecommendationRequestBuilder
     {
         $requestBuilder = new RecommendationRequestBuilder($recommendation);
         $this->setupBuilder($requestBuilder);